proj1a/LinkedListDequeTest.java
import static org.junit.Assert.*;
import org.junit.Test;

/** Performs some basic linked list tests. */
public class LinkedListDequeTest {
	
	/** Utility method for printing out empty checks. */
	public static boolean checkEmpty(boolean expected, boolean actual) {
		if (expected != actual) {
			System.out.println("isEmpty() returned " + actual + ", but expected: " + expected);
			return false;
		}
		return true;
	}

	/* Utility method for printing out empty checks. */
	public static boolean checkSize(int expected, int actual) {
		if (expected != actual) {
			System.out.println("size() returned " + actual + ", but expected: " + expected);
			return false;
		}
		return true;
	}

	/* Prints a nice message based on whether a test passed. 
	 * The \n means newline. */
	public static void printTestStatus(boolean passed) {
		if (passed) {
			System.out.println("Test passed!\n");
		} else {
			System.out.println("Test failed!\n");
		}
	}

	/** Adds a few things to the list,
	 * checking isEmpty() and size() are correct,
	  * finally printing the results. 
	  *
	  * && is the "and" operation. */
	public static void addIsEmptySizeTest() {
		System.out.println("Running add/isEmpty/Size test.");

		LinkedListDeque<String> lld1 = new LinkedListDeque<String>();

		boolean passed = checkEmpty(true, lld1.isEmpty());

		lld1.addFirst("front");
		
		// The && operator is the same as "and" in Python.
		// It's a binary operator that returns true if both arguments true, and false otherwise.
		passed = checkSize(1, lld1.size()) && passed;
		passed = checkEmpty(false, lld1.isEmpty()) && passed;
		lld1.addLast("middle");
		passed = checkSize(2, lld1.size()) && passed;

		lld1.addLast("back");
		passed = checkSize(3, lld1.size()) && passed;

		System.out.println("Printing out deque: ");
		lld1.printDeque();

		printTestStatus(passed);

	}

	/** Adds an item, then removes an item,
	 * and ensures that dll is empty afterwards. */
	public static void addRemoveTest() {

		System.out.println("Running add/remove test.");

		LinkedListDeque<Integer> lld1 = new LinkedListDeque<Integer>();
		// should be empty 
		boolean passed = checkEmpty(true, lld1.isEmpty());

		lld1.addFirst(10);
		// should not be empty 
		passed = checkEmpty(false, lld1.isEmpty()) && passed;

		lld1.removeFirst();
		// should be empty 
		passed = checkEmpty(true, lld1.isEmpty()) && passed;

		printTestStatus(passed);

	}

	@Test
	public void emptySizeTest() {
		LinkedListDeque<Integer> test = new LinkedListDeque<Integer>();
		assertEquals(0, test.size());
		assertTrue(test.isEmpty());
		test.addFirst(1);
		test.removeFirst();
		assertTrue(test.isEmpty());
	}

	@Test
	public void randomAddFirstRemoveTest() {
		LinkedListDeque<Integer> test = new LinkedListDeque<Integer>();
		test.addFirst(0);
		assertEquals(0, (int) test.removeLast());
		test.addFirst(2);
		assertTrue(test.isEmpty());
		assertEquals(2, (int) test.removeLast());
	}

	@Test
	public void testAddAndGetLast() {
		LinkedListDeque<Integer> T = new LinkedListDeque<>();
		T.addFirst(99);
		T.addFirst(100);
		T.addLast(98);
		int A = T.get(1);
		assertEquals(A, 99);
		assertEquals((int) T.getRecursive(2), 98);
		T.printDeque();
	}


	public static void main(String[] args) {
		System.out.println("Running tests.\n");
		addIsEmptySizeTest();
		addRemoveTest();
	}
} 